requirements.txt
schedule==1.2.0
python-dotenv==1.0.0
requests==2.31.0
google-generativeai==0.3.2
flask==3.0.0
gunicorn==21.2.0
playwright==1.40.0